refactor: lift the netlink layer into lnvps_netlink
Increment 7, and a prerequisite for `lvd`. No behaviour change: code
moved, nothing rewritten.

`lnvps_node` owned the only description in the codebase of how to make a
Linux kernel carry a WireGuard data plane. The VPN route-server daemon
needs the same thing -- interfaces, addresses, routes, and WireGuard
itself over netlink rather than by shelling out to `ip` -- and two
answers to that question would drift, with the second one's bugs found
on a machine nobody is watching.

Moved: `NetOps` and its `WgSettings`/`WgObserved` vocabulary, the netlink
`Kernel`, `UnavailableKernel`, the `/proc/sys` helpers, and `netns`,
which `Kernel` opens its socket inside.

Left behind: everything that knows what the data plane is *for*. The
desired-document types, `apply`, `observe` and `DataPlaneState` are a
marketplace node's vocabulary -- bridges, guests, libvirt -- and belong
to the node.

Both are re-exported from their old paths, `net::` and `lnvps_node::netns`,
so no call site moves and the end-to-end harness is untouched. A node
still reads as one thing: everything about its data plane is `net::`,
whether the code lives in this crate or not.

`Kernel` is still namespace-bound, which is right for a node and wrong
for a route server. Giving it a host-namespace constructor is additive
and belongs with the daemon that needs it, not in a refactor.
