+++ lnvps_netlink/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "lnvps_netlink"
+version.workspace = true
+edition.workspace = true
+
+[dependencies]
+anyhow.workspace = true
+async-trait.workspace = true
+ipnetwork.workspace = true
+log.workspace = true
+tokio = { workspace = true, features = ["rt", "net"] }
+
+# Links, addresses and routes.
+rtnetlink = "0.21"
+netlink-packet-route = "0.30"
+futures-util = "0.3.33"
+# The kernel's own WireGuard netlink interface, rather than the `wg` binary.
+defguard_wireguard_rs = "0.11"
+# WireGuard states keys in base64, so parsing one back is part of talking to it.
+base64 = "0.22"
+# Entering a network namespace: setns, and the mounts that pin one.
+nix = { version = "0.30", features = ["sched", "mount", "fs"] }
+
+[dev-dependencies]
+tempfile = "3"
